move pub, adb, and harness_deploy files
