+++ ide/app/lib/services.dart
@@ -7,7 +7,7 @@ library spark.services;
 import 'dart:async';
 import 'dart:isolate';
 
-import 'pub.dart';
+import 'dart/pub.dart';
 import 'services/compiler.dart';
 import 'services/services_common.dart';
 import 'utils.dart';
