+++ ide/app/test/pub_test.dart
@@ -7,7 +7,7 @@ library spark.pub_test;
 import 'package:unittest/unittest.dart';
 
 import 'files_mock.dart';
-import '../lib/pub.dart';
+import '../lib/dart/pub.dart';
 import '../lib/workspace.dart';
 
 defineTests() {
